Use int64_t for "last_write_time" instead of int
Previously, last_write_time was stored as an int, while other time-related
fields (e.g., current_time and next_state_predict_time) use int64_t. This
type mismatch could truncate large simulation timestamps and made time
arithmetic inconsistent.

This commit:
- Changes last_write_time from int to int64_t for refresh bookkeeping
- Aligns last_write_time with the simulator's existing int64_t time fields

+++ initialize.h
@@ -290,7 +290,7 @@ struct blk_info{
     unsigned int free_page_num;        //记录该块中的free页个数，同上
     unsigned int invalid_page_num;     //记录该块中失效页的个数，同上
     int last_write_page;               //记录最近一次写操作执行的页数,-1表示该块没有一页被写过
-    int last_write_time;               //紀錄最近一次寫入該塊的時間，用於 Refresh
+    int64_t last_write_time;               //紀錄最近一次寫入該塊的時間，用於 Refresh
     struct page_info *page_head;       //记录每一子页的状态
 };
